requirements.txt
python-telegram-bot==21.0
openai>=1.0.0
pdfplumber
python-dotenv
aiohttp>=3.9.0
flask